add new tasks

+++ app/src/main/java/pl/edu/pjatk/pro1/ToDoTask.kt
@@ -6,6 +6,6 @@ import java.util.*
 class ToDoTask(
     val name: String,
     val priority: Int,
-    val progress: Float,
+    val progress: Double,
     val deadline: LocalDate
 )
